Separate filters into classes of their own. Move code to compat layer.
