+++ src/core/filters/i-filter.ts
@@ -1,0 +1,4 @@
+export interface IFilter {
+    filterType: string;
+    isFiltered (value): boolean;
+}
